feat(inventory): fix value object

+++ src/main/java/com/go4u/keepitfreshplatform/inventory/domain/model/valueobjects/Quantity.java
@@ -3,7 +3,7 @@ package com.go4u.keepitfreshplatform.inventory.domain.model.valueobjects;
 import jakarta.persistence.Embeddable;
 
 @Embeddable
-public record Quantity(Long quantity) {
+public record Quantity(Integer quantity) {
     public Quantity {
         if (quantity <= 0) {
             throw new IllegalArgumentException("Quantity must be positive.");
